requirements.txt
Flask==2.3.3
pandas==2.0.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0